Event-Driven-Systems/requirements.txt
redis
aioredis
kafka-python
confluent-kafka
asyncio